checkpoint

+++ util/pqueue.h
@@ -36,7 +36,7 @@
 #define PQUEUE_H
 
 /** priority data type */
-typedef unsigned int pqueue_pri_t;
+typedef unsigned long long pqueue_pri_t;
 
 /** callback functions to get/set/compare the priority of an element */
 typedef pqueue_pri_t (*pqueue_get_pri_f)(void *a);
